Move `_get_context` to the helpers module.
